Prep for grapheme-cluster layout clusters
Behavior-neutral groundwork for returning grapheme clusters (rather than
shaping clusters) from layout:

- Add a grapheme-start flag to `CharInfo`, computed in `analyze_text` via
  the ICU4X grapheme segmenter (same mechanism as word/line boundaries).
- Derive `shape_item`'s font-selection segmentation from that flag instead
  of re-running the grapheme segmenter per item, so all downstream
  consumers share one source of truth for grapheme boundaries.
- Widen `ClusterData::text_len` to `u16` (a grapheme cluster can exceed
  255 bytes) and shrink `flags` to `u8` (2 bits used), keeping the struct
  at 24 bytes (now statically asserted).
- Store an emoji flag on `ClusterInfo` computed from ICU4X
  Emoji/Extended_Pictographic/Regional_Indicator properties, replacing the
  hardcoded codepoint ranges in `is_emoji()` (resolves the TODO; regional
  indicators are now classified as emoji).
- Rework the cluster snapshot test renderer to derive per-cluster
  reference text from `Cluster::text_range()` instead of assuming one
  cluster is one `char`.

+++ parley/src/layout/cluster.rs
@@ -448,7 +448,7 @@ impl<'a, B: Brush> Cluster<'a, B> {
     ///
     /// This is only used for tests, and is *not* part of the public API.
     #[doc(hidden)]
-    pub fn text_len(&self) -> u8 {
+    pub fn text_len(&self) -> u16 {
         self.data.text_len
     }
